feat(pointers): Update skia hosts init to unify pointer support
