+++ src/Uno.UI.Runtime.Skia.Wpf/WpfCoreWindowExtension.cs
@@ -28,8 +28,8 @@ namespace Uno.UI.Skia.Platform
 {
 	internal partial class WpfCoreWindowExtension : ICoreWindowExtension
 	{
-		private readonly ICoreWindowEvents _ownerEvents;
 		private readonly WpfHost? _host;
+		private readonly CoreWindow _owner;
 
 		public CoreCursor PointerCursor
 		{
@@ -39,8 +39,7 @@ namespace Uno.UI.Skia.Platform
 
 		public WpfCoreWindowExtension(object owner)
 		{
-			_ownerEvents = (ICoreWindowEvents)owner;
-
+			_owner = (CoreWindow)owner;
 			_host = WpfHost.Current;
 			if (_host is null)
 			{
@@ -60,12 +59,6 @@ namespace Uno.UI.Skia.Platform
 				win.AddHandler(System.Windows.UIElement.KeyDownEvent, (System.Windows.Input.KeyEventHandler)HostOnKeyDown, true);
 			}
 		}
-
-		public void SetPointerCapture(PointerIdentifier pointer)
-			=> WpfHost.Current?.CaptureMouse();
-
-		public void ReleasePointerCapture(PointerIdentifier pointer)
-			=> WpfHost.Current?.ReleaseMouseCapture();
 
 		internal static WpfCanvas? GetOverlayLayer(XamlRoot xamlRoot) =>
 			XamlRootMap.GetHostForRoot(xamlRoot)?.NativeOverlayLayer;
