--- src/Uno.UI/UI/Xaml/Input/Internal/IPointerExtension.cs
@@ -1,11 +1,0 @@
-﻿using Windows.Devices.Input;
-using Windows.UI.Xaml;
-
-namespace Uno.UI.Xaml.Input;
-
-internal interface IPointerExtension
-{
-	void ReleasePointerCapture(PointerIdentifier pointer, XamlRoot xamlRoot);
-
-	void SetPointerCapture(PointerIdentifier pointer, XamlRoot xamlRoot);
-}
