+++ src/Uno.UI/UI/Xaml/Internal/Islands/XamlIslandRoot.cs
@@ -1,4 +1,6 @@
-﻿using Uno.UI.Xaml.Core;
+﻿#nullable enable
+
+using Uno.UI.Xaml.Core;
 using Windows.Foundation;
 using Windows.UI;
 using Windows.UI.Xaml;
