+++ src/Uno.UWP/UI/Core/CoreWindow.skia.cs
@@ -2,6 +2,7 @@
 
 using System;
 using Uno.Foundation;
+using Uno.Foundation.Extensibility;
 using Windows.Foundation;
 
 namespace Windows.UI.Core
@@ -12,19 +13,19 @@ namespace Windows.UI.Core
 		internal event TypedEventHandler<CoreWindow, KeyEventArgs>? NativeKeyUpReceived;
 
 		internal bool IsNativeElement(object content)
-			=> _coreWindowExtension!.IsNativeElement(content);
+			=> _coreWindowExtension.IsNativeElement(content);
 
 		internal void AttachNativeElement(object owner, object content)
-			=> _coreWindowExtension!.AttachNativeElement(owner, content);
+			=> _coreWindowExtension.AttachNativeElement(owner, content);
 
 		internal void DetachNativeElement(object owner, object content)
-			=> _coreWindowExtension!.DetachNativeElement(owner, content);
+			=> _coreWindowExtension.DetachNativeElement(owner, content);
 
 		internal void ArrangeNativeElement(object owner, object content, Rect arrangeRect)
-			=> _coreWindowExtension!.ArrangeNativeElement(owner, content, arrangeRect);
+			=> _coreWindowExtension.ArrangeNativeElement(owner, content, arrangeRect);
 
 		internal Size MeasureNativeElement(object owner, object content, Size size)
-			=> _coreWindowExtension!.MeasureNativeElement(owner, content, size);
+			=> _coreWindowExtension.MeasureNativeElement(owner, content, size);
 
 		void ICoreWindowEvents.RaiseNativeKeyDownReceived(KeyEventArgs args)
 		{
