+++ src/Uno.UWP/UI/Core/ICoreWindowExtension.cs
@@ -3,26 +3,19 @@ using System.Linq;
 using Windows.Devices.Input;
 using Windows.Foundation;
 
-namespace Windows.UI.Core
+namespace Windows.UI.Core;
+
+internal interface ICoreWindowExtension
 {
-	internal interface ICoreWindowExtension
-	{
-		public CoreCursor PointerCursor { get; set; }
-
-		void ReleasePointerCapture(PointerIdentifier pointer);
-
-		void SetPointerCapture(PointerIdentifier pointer);
-
 #if UNO_SUPPORTS_NATIVEHOST
-		bool IsNativeElement(object content);
+	bool IsNativeElement(object content);
 
-		void AttachNativeElement(object owner, object content);
+	void AttachNativeElement(object owner, object content);
 
-		void DetachNativeElement(object owner, object content);
+	void DetachNativeElement(object owner, object content);
 
-		void ArrangeNativeElement(object owner, object content, Rect arrangeRect);
+	void ArrangeNativeElement(object owner, object content, Rect arrangeRect);
 
-		Size MeasureNativeElement(object owner, object content, Size size);
+	Size MeasureNativeElement(object owner, object content, Size size);
 #endif
-	}
 }
