+++ src/Uno.UWP/UI/Input/Preview.Injection/InputInjector.cs
@@ -5,20 +5,20 @@ using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using Windows.Devices.Input;
 using Windows.Foundation;
-using Windows.UI.Core;
 using Uno.Extensions.Specialized;
 
 namespace Windows.UI.Input.Preview.Injection;
 
 public partial class InputInjector
 {
+	[ThreadStatic] private static IInputInjectorTarget? _inputManager;
+
+	internal static void SetTargetForCurrentThread(IInputInjectorTarget manager)
+		=> _inputManager = manager;
+
 	[global::Uno.NotImplemented("__ANDROID__", "__IOS__", "NET461", "__WASM__", "__NETSTD_REFERENCE__", "__MACOS__")]
 	public static InputInjector? TryCreate()
-#if UNO_HAS_MANAGED_POINTERS
-		=> CoreWindow.GetForCurrentThread() is { } window ? new InputInjector(window) : null;
-#else
-		=> null;
-#endif
+		=> _inputManager is not null ? new InputInjector(_inputManager) : null;
 
 	private readonly InjectedInputState _mouse = new(PointerDeviceType.Mouse);
 	private (InjectedInputState state, bool isAdded)? _touch;
@@ -28,18 +28,12 @@ public partial class InputInjector
 	/// </summary>
 	internal InjectedInputState Mouse => _mouse;
 
-#if UNO_HAS_MANAGED_POINTERS
-	private readonly CoreWindow _window;
+	private readonly IInputInjectorTarget _target;
 
-	private InputInjector(CoreWindow window)
+	private InputInjector(IInputInjectorTarget target)
 	{
-		_window = window;
+		_target = target;
 	}
-#else
-	private InputInjector()
-	{
-	}
-#endif
 
 	[global::Uno.NotImplemented("__ANDROID__", "__IOS__", "NET461", "__WASM__", "__NETSTD_REFERENCE__", "__MACOS__")]
 	public void InitializeTouchInjection(InjectedInputVisualizationMode visualMode)
@@ -56,7 +50,7 @@ public partial class InputInjector
 		{
 			var cancel = new InjectedInputTouchInfo { PointerInfo = new() { PointerOptions = InjectedInputPointerOptions.Canceled } };
 
-			InjectPointerRemoved(cancel.ToEventArgs(_touch.Value.state));
+			_target.InjectPointerRemoved(cancel.ToEventArgs(_touch.Value.state));
 
 			_touch = null;
 		}
@@ -77,17 +71,17 @@ public partial class InputInjector
 
 			if (_touch is { isAdded: false })
 			{
-				InjectPointerAdded(args);
+				_target.InjectPointerAdded(args);
 				_touch = (touch, isAdded: true);
 			}
 
 			touch.Update(args);
 
-			InjectPointerUpdated(args);
+			_target.InjectPointerUpdated(args);
 
 			if (info.PointerInfo.PointerOptions.HasFlag(InjectedInputPointerOptions.PointerUp))
 			{
-				InjectPointerRemoved(args);
+				_target.InjectPointerRemoved(args);
 				_touch = (touch, isAdded: false);
 			}
 		}
@@ -101,33 +95,7 @@ public partial class InputInjector
 			var args = info.ToEventArgs(_mouse!);
 			_mouse!.Update(args);
 
-			InjectPointerUpdated(args);
+			_target.InjectPointerUpdated(args);
 		}
 	}
-
-#if UNO_HAS_MANAGED_POINTERS
-	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	private void InjectPointerAdded(PointerEventArgs args)
-		=> _window.InjectPointerAdded(args);
-
-	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	private void InjectPointerUpdated(PointerEventArgs args)
-		=> _window.InjectPointerUpdated(args);
-
-	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	private void InjectPointerRemoved(PointerEventArgs args)
-		=> _window.InjectPointerRemoved(args);
-#else
-	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	private void InjectPointerAdded(PointerEventArgs args)
-		=> throw new InvalidOperationException("Input injection is not supported on this platform.");
-
-	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	private void InjectPointerUpdated(PointerEventArgs args)
-		=> throw new InvalidOperationException("Input injection is not supported on this platform.");
-
-	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	private void InjectPointerRemoved(PointerEventArgs args)
-		=> throw new InvalidOperationException("Input injection is not supported on this platform.");
-#endif
 }
